requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
Werkzeug==3.0.1
python-dotenv==1.0.0
mysqlclient==2.2.1
PyMySQL==1.1.0
